Move the swiftui chart logic to dydxLineChartViewModel for reuse
