+++ dydx/dydxViews/dydxViews/_v4/Vault/Landing/dydxVaultChartViewModel.swift
@@ -15,15 +15,6 @@ import dydxChart
 import dydxFormatter
 
 public class dydxVaultChartViewModel: PlatformViewModel {
-    public struct Entry {
-        let date: Double
-        let value: Double
-
-        public init(date: Double, value: Double) {
-            self.date = date
-            self.value = value
-        }
-    }
 
     @Published public var selectedValueType: ValueTypeOption = .pnl
     @Published public var selectedValueTime: ValueTimeOption = .sevenDays
@@ -31,11 +22,7 @@ public class dydxVaultChartViewModel: PlatformViewModel {
     fileprivate let valueTypeOptions = ValueTypeOption.allCases
     fileprivate let valueTimeOptions = ValueTimeOption.allCases
 
-    @Published public var entries: [Entry] = []
-    fileprivate var isPositive: Bool { (entries.last?.value ?? -Double.infinity) >= (entries.first?.value ?? -Double.infinity) }
-    fileprivate var lineColor: Color { isPositive ? ThemeSettings.positiveColor.color : ThemeSettings.negativeColor.color }
-    fileprivate var datesDomain: ClosedRange<Double> { (entries.map(\.date).min() ?? 0)...(entries.map(\.date).max() ?? 0) }
-    fileprivate var valuesDomain: ClosedRange<Double> { (entries.map(\.value).min() ?? 0)...(entries.map(\.value).max() ?? 0) }
+    @Published public var chart = dydxLineChartViewModel()
 
     public enum ValueTypeOption: CaseIterable, RadioButtonContentDisplayable {
         case pnl
@@ -83,11 +70,6 @@ public class dydxVaultChartViewModel: PlatformViewModel {
 private struct dydxVaultChartView: View {
     @ObservedObject var viewModel: dydxVaultChartViewModel
 
-    private var chartGradient: Gradient {
-        Gradient(colors: [viewModel.lineColor.opacity(0.25),
-                          viewModel.lineColor.opacity(0)])
-    }
-
     private var radioButtonsRow: some View {
         GeometryReader { geometry in
             ScrollView(.horizontal, showsIndicators: false) {
@@ -115,43 +97,10 @@ private struct dydxVaultChartView: View {
         .frame(height: 38)
     }
 
-    private var chart: some View {
-        Chart(viewModel.entries, id: \.date) { entry in
-            LineMark(x: .value("", entry.date),
-                     y: .value("", entry.value))
-            .lineStyle(StrokeStyle(lineWidth: 2))
-            .foregroundStyle(viewModel.lineColor.gradient)
-            .interpolationMethod(.monotone)
-            .symbolSize(0)
-            // adds gradient shading
-            AreaMark(
-                x: .value("", entry.date),
-                yStart: .value("", viewModel.valuesDomain.lowerBound),
-                yEnd: .value("", entry.value)
-            )
-            .foregroundStyle(chartGradient)
-        }
-        .chartXAxis(.hidden)
-        .chartYAxis {
-            AxisMarks(values: .automatic) {
-                let value = dydxFormatter.shared.condensedDollar(number: $0.as(Double.self), digits: 1)
-                AxisValueLabel {
-                    if let value {
-                        Text(value)
-                    }
-                }
-                .font(Font.system(size: 8))
-                .foregroundStyle(ThemeColor.SemanticColor.textTertiary.color)
-            }
-        }
-        .chartXScale(domain: .automatic(includesZero: false))
-        .chartYScale(domain: .automatic(includesZero: false))
-    }
-
     var body: some View {
         VStack(spacing: 8) {
             radioButtonsRow
-            chart
+            viewModel.chart.createView()
         }
     }
 }
